+++ Script/test_linux.sh
@@ -14,18 +14,18 @@ if [ -n "$1" ]; then
 	PROJECT_NAME=$1
 fi
 
-if [ ! -f /opt/${PROJECT_NAME}/share/applications/org.Rabbit.${PROJECT_NAME}.desktop ]; then
-	echo "There are not /opt/share/applications/org.Rabbit.${PROJECT_NAME}.desktop"
+if [ ! -f /opt/${PROJECT_NAME}/share/applications/io.github.KangLin.SerialPortAssistant.desktop ]; then
+	echo "There are not /opt/share/applications/io.github.KangLin.SerialPortAssistant.desktop"
 	EXIT_CODE=$[EXIT_CODE+1]
 fi
 
-if [ ! -f /usr/share/applications/org.Rabbit.${PROJECT_NAME}.desktop ]; then
-	echo "There are not /usr/share/applications/org.Rabbit.${PROJECT_NAME}.desktop"
+if [ ! -f /usr/share/applications/io.github.KangLin.SerialPortAssistant.desktop ]; then
+	echo "There are not /usr/share/applications/io.github.KangLin.SerialPortAssistant.desktop"
 	EXIT_CODE=$[EXIT_CODE+1]
 fi
 
-if [ ! -f /usr/share/pixmaps/org.Rabbit.${PROJECT_NAME}.png ]; then
-    echo "There are not /usr/share/pixmaps/org.Rabbit.${PROJECT_NAME}.png"
+if [ ! -f /usr/share/pixmaps/io.github.KangLin.SerialPortAssistant.png ]; then
+    echo "There are not /usr/share/pixmaps/io.github.KangLin.SerialPortAssistant.png"
     EXIT_CODE=$[EXIT_CODE+1]
 fi